+++ src/zip/zfile.c
@@ -20,7 +20,7 @@ static unzFile unzfile = NULL;
 static char basedir[MAX_PATH];
 static char *basedirend;
 static char zip_cache[4096];
-static int  zip_cached_len;
+static size_t  zip_cached_len;
 static int  zip_filepos;
 
 
@@ -154,7 +154,7 @@ int zclose(int fd)
 	ZIPファイル内のファイルを読み込む
 ------------------------------------------------------*/
 
-int zread(int fd, void *buf, unsigned size)
+size_t zread(int fd, void *buf, size_t size)
 {
 	if (unzfile == NULL)
 		return read(fd, buf, size);
@@ -187,13 +187,13 @@ int zgetc(int fd)
 	ZIPファイル内のファイルのサイズを取得
 ------------------------------------------------------*/
 
-int zsize(int fd)
+size_t zsize(int fd)
 {
 	unz_file_info info;
 
 	if (unzfile == NULL)
 	{
-		int len, pos = lseek(fd, 0, SEEK_CUR);
+        off_t len, pos = lseek(fd, 0, SEEK_CUR);
 
 		len = lseek(fd, 0, SEEK_END);
 		lseek(fd, pos, SEEK_CUR);
